# samples/openapi/openapi.yaml
...
openapi: 3.1.2
info:
  title: "Global"
  version: "1.0.0"
paths:
...
  "/user":
    post:
      operationId: "AdminUserCreate"
      summary: "Create"
      tags:
        - "Admin.User"
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "No Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdUser"
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdUser"
    get:
      operationId: "AdminUserList"
      summary: "List"
      tags:
        - "Admin.User"
      security:
        - {}
      parameters:
        -
          name: "text"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "No Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultsIdentifiedUser"
  "/user/{user_id}":
    delete:
      operationId: "AdminUserDelete"
      summary: "Delete"
      tags:
        - "Admin.User"
      security:
        - {}
      parameters:
        -
          name: "user_id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "No Error"
        "204":
          description: "No Content"
        "205":
          description: "Reset Content"
    get:
      operationId: "AdminUserGet"
      summary: "Get"
      tags:
        - "Admin.User"
      security:
        - {}
      parameters:
        -
          name: "user_id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "No Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifiedUser"
    patch:
      operationId: "AdminUserPatch"
      summary: "Patch"
      tags:
        - "Admin.User"
      security:
        - {}
      parameters:
        -
          name: "user_id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedUser"
      responses:
        "200":
          description: "No Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdUser"
    put:
      operationId: "AdminUserUpdate"
      summary: "Update"
      tags:
        - "Admin.User"
      security:
        - {}
      parameters:
        -
          name: "user_id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "No Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdUser"
components:
  schemas:
...
    User:
      type: "object"
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"
        username:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        hire_date:
          type: "string"
          format: "date-time"
        birth_date:
          type: "string"
          format: "date-time"
        picture:
          type: "string"
          format: "uri"
      required:
        - "first_name"
        - "last_name"
        - "username"
    IdentifiedUser:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        first_name:
          type: "string"
        last_name:
          type: "string"
        username:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        hire_date:
          type: "string"
          format: "date-time"
        birth_date:
          type: "string"
          format: "date-time"
        picture:
          type: "string"
          format: "uri"
        creation_date:
          type: "string"
          format: "date-time"
        last_update_date:
          type: "string"
          format: "date-time"
      required:
        - "id"
        - "first_name"
        - "last_name"
        - "username"
        - "creation_date"
        - "last_update_date"
    IdUser:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
      required:
        - "id"
    PatchedUser:
      type: "object"
      properties:
        first_name:
          $ref: "#/components/schemas/PatchableValueValue"
        last_name:
          $ref: "#/components/schemas/PatchableValueValue"
        username:
          $ref: "#/components/schemas/PatchableValueValue"
        email:
          $ref: "#/components/schemas/PatchableValueValue"
        phone:
          $ref: "#/components/schemas/PatchableValueValue"
        hire_date:
          $ref: "#/components/schemas/PatchableValueValue"
        birth_date:
          $ref: "#/components/schemas/PatchableValueValue"
        picture:
          $ref: "#/components/schemas/PatchableValueValue"
      required:
        - "first_name"
        - "last_name"
        - "username"
    PatchableValueValue:
      type: "object"
    PagedResultsIdentifiedUser:
      type: "object"
...
  securitySchemes:
    apiKey:
      type: "apiKey"
      in: "header"
      name: "Authorization"